\name{Recreation}
\alias{Recreation}
\docType{data}
\title{Recreation Data}
\description{
Ten subjects were asked to judge differences, in terms of his/her own preference, between 15 forms of recreations. Each particular subject judged every one of 105 pairs of forms of recreation in a dissimilarity scale of 25 categories (numbered from 1 to 25). In this data set, only the lower right triangle excluding the diagonal was originally considered for each dissimilarity matrix because the natural asymmetry in judements. Recreations forms were Concert (Ct), Museum (M), Theater  (T), Movie (Mv), Television (TV), Conference (Cf), Reading (R), Hockey (H), Ballet (B), Debate (D), Fashion (F), Doc-film (Df), Exhibition (E), Shopping (S) and Restaurant (Re).
}
\usage{data(Miller)}
\format{
  A list of dissimilarity matrices.
}

\references{
Ramsay, J. O. (1978). MULTISCALE: Four programs for multidimensional scaling by the method of maximum likelihood. Chicago: National Educational Resources, Inc. 
}
\examples{
data(Recreation)
}
\keyword{datasets}
