\name{Wang}
\alias{Wang}
\docType{data}
\title{Consonant Confusions in Noise}
\description{
In this data set by Wang and Bilger (1973), the same consonant phonemes considered as in Miller & Nicely (1955), except that "m"" and "n"" are substituted for "tS" as in "cheap", and "d3" as in "jeep"", were combined with each of the three vowels /i,a,u/ under six speach-to-noise ratios. To obtain dissimilarities, the normalization approach proposed in Hubert (1972) was applied. 
}
\usage{data(Wang)}
\format{
  An asymmetric dissimilarity matrix. 
}

\references{
Wang, M. D., & Bilger, R. C. (1973). Consonant confusions in noise: A study of perceptual features. The Journal of the Acoustical Society of America, 54(5), 1248--1266.

Miller, G. A., & Nicely, P. E. (1955). An analysis of perceptual confusions among some English consonants. The Journal of the Acoustic Society of America, 27, 338--352.

Hubert, L. (1972). Some extensions of Johnson's hierarchical clustering algorithms. Psychometrika, 37, 261--274.
}

\seealso{\code{\link{Miller}}}


\examples{
data(Wang)
}
\keyword{datasets}
